feat(textinput): allow separate import for react and css
