+++ packages/textinput/react.d.ts
@@ -1,0 +1,6 @@
+export {
+  default,
+  TextInputStatics,
+  TextInputProps
+} from './dist/esm/react/index'
+export * from './dist/esm/vars/index'
