+++ packages/textinput/react.mjs
@@ -1,0 +1,4 @@
+import TextInput from './dist/esm/react/index'
+
+export default TextInput
+export * from './dist/esm/vars/index'
