+++ packages/textinput/src/index.ts
@@ -1,2 +1,8 @@
-export { default, TextInputProps } from './react/index'
+import '@pluralsight/ps-design-system-halo/styles.css'
+import '@pluralsight/ps-design-system-icon/styles.css'
+
+import './css/index.css'
+import TextInput from './react/index'
+
+export default TextInput
 export * from './vars/index'
