+++ packages/textinput/src/react/index.tsx
@@ -1,5 +1,5 @@
-import Halo from '@pluralsight/ps-design-system-halo'
-import { WarningIcon } from '@pluralsight/ps-design-system-icon'
+import Halo from '@pluralsight/ps-design-system-halo/react'
+import { WarningIcon } from '@pluralsight/ps-design-system-icon/react'
 import {
   useTheme,
   names as themeNames
@@ -11,7 +11,6 @@ import {
 } from '@pluralsight/ps-design-system-util'
 import React from 'react'
 
-import '../css/index.css'
 import * as vars from '../vars/index'
 
 const styles = {
