+++ packages/textinput/styles.css
@@ -1,0 +1,4 @@
+@import '@pluralsight/ps-design-system-halo/styles.css';
+@import '@pluralsight/ps-design-system-icon/styles.css';
+
+@import './dist/esm/css/index.css';
